+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/EditWithGroups.tsx
@@ -7,7 +7,7 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { Language, LocalizedFormProperties } from "@lib/types/form-builder-types";
 import { ElementPanel } from ".";
 import { ConfirmationDescriptionWithGroups } from "./ConfirmationDescriptionWithGroups";
-import { RichTextLockedWithGroups } from "./elements/RichTextLockedWithGroups";
+import { RichTextLocked } from "./elements/RichTextLocked";
 import { ExpandingInput } from "@formBuilder/components/shared/ExpandingInput";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
 import { useRehydrate } from "@lib/store/hooks/useRehydrate";
@@ -142,7 +142,8 @@ export const EditWithGroups = ({ id, locale }: { id: string; locale: string }) =
       </div>
       {/* Form Intro + Title Panel */}
       {groupId === "start" && (
-        <RichTextLockedWithGroups
+        <RichTextLocked
+          maxLength={20000}
           hydrated={hasHydrated}
           className="rounded-t-lg"
           summaryText={t("startFormIntro")}
@@ -180,7 +181,8 @@ export const EditWithGroups = ({ id, locale }: { id: string; locale: string }) =
       )}
       {/* Privacy Panel */}
       {groupId === "start" && (
-        <RichTextLockedWithGroups
+        <RichTextLocked
+          maxLength={50000}
           beforeContent={<PrivacyDescriptionBefore />}
           summaryText={t("groups.privacy.summary")}
           detailsText={
@@ -208,7 +210,8 @@ export const EditWithGroups = ({ id, locale }: { id: string; locale: string }) =
       </div>
       {/* Confirmation*/}
       {groupId === "end" && (
-        <RichTextLockedWithGroups
+        <RichTextLocked
+          maxLength={20000}
           summaryText={t("groups.confirmation.summary")}
           beforeContent={
             <div>