+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/branding/components/options.ts
@@ -269,4 +269,22 @@ export const options = [
     logoTitleEn: "The Future of Sport in Canada Commission",
     logoTitleFr: "Commission sur l’avenir du sport au Canada",
   },
+  {
+    name: "cas",
+    urlEn: "https://www.cas-satj.gc.ca/en/home",
+    urlFr: "https://www.cas-satj.gc.ca/fr/accueil",
+    logoEn: "/img/branding/cas-en.svg",
+    logoFr: "/img/branding/cas-fr.svg",
+    logoTitleEn: "Courts Administration Service",
+    logoTitleFr: "Service administratif des tribunaux judiciaires",
+  },
+  {
+    name: "pedc",
+    urlEn: "https://www.canada.ca/en/prairies-economic-development.html",
+    urlFr: "https://www.canada.ca/fr/developpement-economique-prairies.html",
+    logoEn: "/img/branding/pedc-en.svg",
+    logoFr: "/img/branding/pedc-fr.svg",
+    logoTitleEn: "Prairies Economic Development Canada",
+    logoTitleFr: "Développement économique Canada pour les Prairies",
+  },
 ];