+++ components/clientComponents/forms/Form/SubmitButton.tsx
@@ -5,29 +5,29 @@ import { cn } from "@lib/utils";
 import { useTranslation } from "@i18n/client";
 import { logMessage } from "@lib/logger";
 import { useFeatureFlags } from "@lib/hooks/useFeatureFlags";
+import { ga } from "@lib/client/clientHelpers";
 
 interface SubmitButtonProps {
   getFormDelay: () => number;
   formID: string;
   formTitle: string;
   disabled: boolean;
-  status?: string;
+  submissionError?: boolean;
 }
 export const SubmitButton: React.FC<SubmitButtonProps> = ({
   getFormDelay,
   formID,
   formTitle,
   disabled,
-  status,
+  submissionError,
 }) => {
   const { t } = useTranslation();
   const [formTimerState, { startTimer, checkTimer, disableTimer }] = useFormTimer();
   const [submitTooEarly, setSubmitTooEarly] = useState(false);
   const screenReaderRemainingTime = useRef(formTimerState.remainingTime);
   const formDelay = useRef(getFormDelay());
-
-  // Used to show a spinner for the initial hCAPTCHA verification step
-  const [submitting, setSubmitting] = useState(false);
+  // Used to show a spinner for the initial validation and hCAPTCHA loading steps
+  const [loading, setLoading] = useState(false);
 
   const { getFlag } = useFeatureFlags();
   const timerEnabled = getFlag("formTimer");
@@ -62,6 +62,12 @@ export const SubmitButton: React.FC<SubmitButtonProps> = ({
       };
     }
   }, [checkTimer, formTimerState.remainingTime, formTimerEnabled]);
+
+  useEffect(() => {
+    if (submissionError) {
+      setLoading(false);
+    }
+  }, [submissionError]);
 
   return (
     <>
@@ -106,9 +112,7 @@ export const SubmitButton: React.FC<SubmitButtonProps> = ({
           screenReaderRemainingTime.current = formTimerState.remainingTime;
           if (!formTimerState.canSubmit) {
             e.preventDefault();
-            window.dataLayer = window.dataLayer || [];
-            window.dataLayer.push({
-              event: "form_submission_spam_trigger",
+            ga("form_submission_spam_trigger", {
               formID: formID,
               formTitle: formTitle,
               submitTime: formTimerState.remainingTime,
@@ -120,10 +124,9 @@ export const SubmitButton: React.FC<SubmitButtonProps> = ({
           }
           // Only change state if submitTooEarly is already set to true
           submitTooEarly && setSubmitTooEarly(false);
-
-          setSubmitting(true);
+          setLoading(true);
         }}
-        loading={submitting && status !== "Error"}
+        loading={loading}
       >
         {t("submitButton")}
       </Button>